completed first step of nondeterminism

+++ src/aiModels/PotentialDestination.java
@@ -3,17 +3,17 @@ package aiModels;
 import primary.Point;
 
 public class PotentialDestination {
-	public double expectedArrivalRate;
+	public int expectedArrivalRate;
 	public int targetX;
 	public int targetY;
 	
-	public PotentialDestination(double newRate, int newX, int newY) {
+	public PotentialDestination(int newRate, int newX, int newY) {
 		expectedArrivalRate = newRate;
 		targetX = newX;
 		targetY = newY;
 	}
 	
-	public PotentialDestination(double newRate, Point targetP) {
+	public PotentialDestination(int newRate, Point targetP) {
 		expectedArrivalRate = newRate;
 		targetX = targetP.x;
 		targetY = targetP.y;
